rccl: fix possible int underflow in p2p_resiliency_recovery_ainic.cc

+++ projects/rccl/src/transport/net_ib_cast/p2p_resiliency_recovery_ainic.cc
@@ -211,7 +211,7 @@ ncclResult_t IbCastPortRecoveryPostAckWithQpnsAinic(struct ncclIbPortRecoveryCon
 void IbCastPortRecoveryExtractQpnsAinic(struct ncclIbPortRecoveryContext* recoveryContext, struct ibv_wc* wc) {
   if (!IbCastAinicRoce) return;
   struct ncclIbResiliencyDev* resDev = &recoveryContext->resCtx->devs[recoveryContext->devIndex];
-  int payloadLen = wc->byte_len - NCCL_IB_UD_GRH_SIZE;
+  int payloadLen = (int)(wc->byte_len - NCCL_IB_UD_GRH_SIZE);
   if (payloadLen <= 0) {
     recoveryContext->nRemoteQpns = 0;
     return;
